requirements.txt
setuptools~=60.10.0
pymongo~=4.3.3
pandas~=1.4.4
sensor~=0.0.1
-e .

numpy~=1.24.2
python-dotenv~=1.0.0